Create Description.md
A desktop application written in Python and Qt. The application will also run on all major operating systems. The basic Requirement 
of the application is to be able to Download files from the Web. The application will actually download the file from the URL, the user 
specifies in a QLineEdit box. The QProgressBar will also be present in our application, and it will show the user the progress, the 
percentage of the file that has been downloaded so far. It allows the user to browse their computer, so that they can specify the location 
where they want to save their file. When the download completes, we should notify the user, by showing a message on the screen that says, download completed, or something similar.